feat(validation): validation or delete

+++ src/Modules/spending/UseCases/DeleteSpending/deleteSpendingService.js
@@ -1,14 +1,21 @@
 const { supabase } = require("../../../../database/Supabase");
 
 class DeleteSpendingService {
-  async deleteSpending(id) {
+  async deleteSpending(id, uuid) {
     const { data: spending, error } = await supabase
       .from("spending")
       .delete()
+      .eq("user_id", uuid)
       .match({ id });
+    console.log("error", error);
+    console.log("spending", spendings);
 
-    console.log(spending);
-    return spending;
+    const data = {
+      spending,
+      error,
+    };
+
+    return data;
   }
 }
